// src/app/languages/interface/cartmain/cartmain.interface.ts
export interface ICartMain {
    viewMore: string;
    wantTogetFreeoffer: string;
    doyouwanttocontinue: string;
    youhaveRejected: string;
    wantToProceed: string;
    clickhere: string;
    note: string;
    deliver: string;
    items: string;
    proceed: string;
    yes: string;
    no: string;
}